Fix Phase-2 resume critic blockers: skip safety, bigint seeds, credited upload
Require on-disk segment coverage before skipping, clear sidecars when resume
is disabled, migrate torrent seed byte columns to bigint, and persist credited
upload so SeedRatioLimit survives process restart.

+++ internal/db/migrations/0019_download_resume_seed_state.sql
@@ -3,6 +3,7 @@
 -- Reason: Phase-2 ARR-parity — staging sidecar is SoT for segments; DB mirror is
 --   for UI/debug. Seed ratio/duration windows were in-memory only and reset on
 --   every process restart, so seed limits never accumulated across boots.
+-- Note: seed_baseline_up stores credited upload bytes (not the live handle counter).
 -- Troubleshooting: usenet_resume_state / torrent_seed_state rows keyed by download_gid;
 --   cleared on import (usenet) and seed-stop/cancel (torrent).
 -- Review if: resume state is folded into grabs or seed state gains per-file paths.
@@ -16,8 +17,8 @@ CREATE TABLE IF NOT EXISTS usenet_resume_state (
 CREATE TABLE IF NOT EXISTS torrent_seed_state (
     download_gid      TEXT PRIMARY KEY NOT NULL,
     seed_started_at   TEXT NOT NULL DEFAULT '',
-    seed_baseline_up  INTEGER NOT NULL DEFAULT 0,
-    seed_total_bytes  INTEGER NOT NULL DEFAULT 0,
+    seed_baseline_up  bigint NOT NULL DEFAULT 0,
+    seed_total_bytes  bigint NOT NULL DEFAULT 0,
     updated_at        TEXT NOT NULL DEFAULT ''
 );
